Remove casts from DUCK video decoder which cause function cast warnings
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@3686 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/libs/video/dukvid.c
@@ -727,7 +727,8 @@ dukv_SeekTime (THIS_PTR, float time)
 	//TFB_DuckVideoDecoder* dukv = (TFB_DuckVideoDecoder*) This;
 	uint32 frame = (uint32) (time * DUCK_GENERAL_FPS);
 	
-	return (float) dukv_SeekFrame (This, frame) / DUCK_GENERAL_FPS;
+	// Note that DUCK_GENERAL_FPS is a float constant
+	return dukv_SeekFrame (This, frame) / DUCK_GENERAL_FPS;
 }
 
 static uint32
